+++ generators/server/index.js
@@ -88,6 +88,7 @@ module.exports = class ServerGenerator extends EngageGenerator {
       }
 
       json.scripts.dev = `nodemon ${serverFileName}`;
+      json.scripts["dev:debug"] = "nodemon  -- --inspect app.js";
     });
 
     this._addDependencies("nodemon", null, { dev: true });